+++ BackEnd/SOA/src/main/java/com/moham/coursemores/repository/HashtagOfCourseRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface HashtagOfCourseRepository extends JpaRepository<HashtagOfCourse, Integer> {
-    List<HashtagOfCourse> findByCourseId(int courseId);
-    void deleteByCourseId(int courseId);
+public interface HashtagOfCourseRepository extends JpaRepository<HashtagOfCourse, Long> {
+    List<HashtagOfCourse> findByCourseId(Long courseId);
+    void deleteByCourseId(Long courseId);
 }
