+++ BackEnd/SOA/src/main/java/com/moham/coursemores/repository/RegionRepository.java
@@ -3,5 +3,5 @@ package com.moham.coursemores.repository;
 import com.moham.coursemores.domain.Region;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RegionRepository extends JpaRepository<Region, Integer> {
+public interface RegionRepository extends JpaRepository<Region, Long> {
 }
